bcachefs: Pad degraded extents with BCH_SB_MEMBER_INVALID pointers
Ensure that degraded extents are accounted as degraded, and display
correctly in 'bcachefs fs usage'.

When we write an extent with less durability than the replicas setting,
add extra pointers to BCH_SB_MEMBER_INVALID in
bch2_bkey_set_needs_rebalance().

+++ fs/bcachefs/btree/types.h
@@ -560,7 +560,7 @@ struct btree_trans {
 	struct bch_fs_usage_base fs_usage_delta;
 
 	unsigned		journal_u64s;
-	unsigned		extra_disk_res; /* XXX kill */
+	u64			extra_disk_res;
 
 	__BKEY_PADDED(btree_path_down, BKEY_BTREE_PTR_VAL_U64s_MAX);
